--- gl/basic/input
@@ -1,7 +1,0 @@
-0.13454 0.1344535
-0.611110 0.1011
-0.4545 0.134534
-0.243545 -0.3345
-0.54534 -0.8345435
--0.43454 0.434535
-0.5345435 0.5345535
