+++ gl/basic/modern-tutorials-master/tut05_cube/Makefile
@@ -1,6 +1,7 @@
 LDLIBS=-lglut -lGLEW -lGL -lm
-all: cube
+all: cube line_cube
 clean:
 	rm -f *.o cube
 cube: ../common/shader_utils.o
+line_cube: ../common/shader_utils.o
 .PHONY: all clean
