+++ gl/surface/create_knots.c
@@ -1,0 +1,10 @@
+#include<stdio.h>
+#include<time.h>
+#include<stdlib.h>
+#define M m+1
+#define N n+1
+int m
+void create_knots()
+{
+     
+}
