+++ gl/surface/demo/Makefile
@@ -1,0 +1,5 @@
+LDLIBS=-lGLEW -lGL -lGLU -lglut  
+surface: Vec3.h main.cpp
+	g++ main.cpp -o surface $(LDLIBS)
+clean:
+	rm *~ *.o
