--- gl/surface/input.txt~
@@ -1,40 +1,0 @@
-6 6
-2 2
-0 0 0 0.2 0.5 0.7 1 1 1
-0 0 0 0.2 0.5 0.7 1 1 1
--30 -36 -30
--30 -26 -20
--30 -20 -10
--30 -18 0
--30 -20 10
--30 -26 20
--20 -26 -30
--20 -16 -20
--20 -10 -10
--20 -8 0
--20 -10 10
--20 -16 20
--10 -20 -30
--10 -10 -20
--10 -4 -10
--10 -2 0
--10 -4 10
--10 -10 20
-0 -18 -30
-0 -8 -20
-0 -2 -10
-0 -0 0
-0 -2 10
-0 -8 20
-10 -20 -30
-10 -10 -20
-10 -4 -10
-10 -2 0
-10 -4 10
-10 -10 20
-20 -26 -30
-20 -16 -20
-20 -10 -10
-20 -8 0
-20 -10 10
-20 -16 20
